requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
tqdm==4.65.0
mlflow==2.7.1
pytest==7.4.0
